updated to test for bind-utils and dos2unix
Minimal CentOS install is really that, minimal.  We need a few elements
for the script to run a bit more cleanly and humanely.
We need to defend against the ^M character as some may save the config
in notepad and SCP it to the host in question
We also will need the host command from bin-utils to determine FQDN,
otherwise we get an error state on the use of the installer.